Rename wrapper to container
